requirementshttp.txt

mysql.connector
